Improve darkening and lightening of color shades
If a grid cell's color has been previously darkened with the darken
brush, using the lighten brush on it will first bring it back to its
original shade, before then starting to lighten it to be completely
white. The same works the other way around, for darkening a previously
lightened color.

Also improved the way shading calculations are made by rounding up
the results of the calculations with Math.ceil(). This makes sure
that each shading step will always advance at least by a value of one,
where previously it would not, as a calculation could result in a
decimal value lower than one, equaling to a change of zero, with the
way the calculation and assignment was previously done.