Add content and logic to clear the grid
Added a "Clear Grid" button that, when clicked, will remove the color
of every grid cell. This creates a faster flow of being able to start
over without needing to create a new grid of the same size. Also
made a style change so that a button that will not stay selected,
such as the "grid buttons", will now have an active state to improve
the user experience, letting the user know that the button has been
clicked.